requirements.txt
fastapi[standard]
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
python-multipart==0.0.9
jinja2==3.1.4
itsdangerous==2.2.0
PyJWT